// src/shared/enums.ts
export enum ListTypeEnum {
  PERSON,
  PROFESSION,
  PERSON_FROM_PROFESSION
};

export enum PersonEnum {
  ID = 'id',
  NAME = 'name',
  THUMBNAIL = 'thumbnail',
  AGE = 'age',
  WEIGHT = 'weight',
  HEIGHT = 'height',
  HAIR_COLOR = 'hair_color',
  PROFESSION = 'professions',
  FRIENDS = 'friends',
}